fix api request form

+++ src/main/java/fpt/capstone/vuondau/moodle/request/GetMoodleAssignmentIdsCourseRequest.java
@@ -5,13 +5,13 @@ import java.util.List;
 
 public class GetMoodleAssignmentIdsCourseRequest {
 
-    private List<Integer> assignmentids ;
+    private List<Long> assignmentids ;
 
-    public List<Integer> getAssignmentids() {
+    public List<Long> getAssignmentids() {
         return assignmentids;
     }
 
-    public void setAssignmentids(List<Integer> assignmentids) {
+    public void setAssignmentids(List<Long> assignmentids) {
         this.assignmentids = assignmentids;
     }
 }
